chore(mux-uploader): improve element modularity (#539)
* Update dev watch scripts

* Break progress indicator out to own component

* Refactoring module loading

* Remove old assertion

* Rexport type for react package

* Consolidate style

* Bump build

* use globalThis

* Rename progress-indicator to progress

* Update template

* Fix test

* Update constants

* Address Christian's feedback

* Include Element in export name
