+++ packages/mux-uploader/src/constants.ts
@@ -1,0 +1,9 @@
+export type ProgressTypes = {
+  BAR: 'bar';
+  RADIAL: 'radial';
+};
+
+export const ProgressTypes: ProgressTypes = {
+  BAR: 'bar',
+  RADIAL: 'radial',
+};
