+++ packages/mux-uploader/test/uploader.test.js
@@ -12,7 +12,6 @@ describe('<mux-uploader>', () => {
       status
     ></mux-uploader>`);
 
-    assert.equal(uploader.getAttribute('type'), 'bar', 'default progress type is reflected');
     assert.equal(
       uploader.getAttribute('endpoint'),
       'https://my-authenticated-url/storage?your-url-params',
